# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/hitl/{requestId}/approve": {
      "post": {
        "operationId": "HitlController_approve",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveAndExecuteResult_Output"
                }
              }
            }
          }
        },
        "summary": "Aprueba una acción pendiente — ejecuta la tool real igual que /approve en Telegram",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/{requestId}/reject": {
      "post": {
        "operationId": "HitlController_reject",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResultDto_Output"
                }
              }
            }
          }
        },
        "summary": "Rechaza una acción pendiente",
        "tags": [
          "hitl"
        ]
      }
    },
...
  },
  "info": {
    "title": "Jin Core API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ResolveAndExecuteResult_Output": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "outcome": {
                "type": "string",
                "enum": [
                  "awaiting-second"
                ]
              }
            },
            "required": [
              "outcome"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "outcome": {
                "type": "string",
                "enum": [
                  "resolved"
                ]
              },
              "toolName": {
                "type": "string"
              },
              "result": {}
            },
            "required": [
              "outcome",
              "toolName",
              "result"
            ],
            "additionalProperties": false
          }
        ]
      },
      "OkResultDto_Output": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
...
    }
  }
}